+++ instrumentation/ethon/lib/opentelemetry/instrumentation/ethon/patches/stable/easy.rb
@@ -12,20 +12,11 @@ module OpenTelemetry
         module Stable
           # Ethon::Easy patch for instrumentation
           module Easy
-            ACTION_NAMES_TO_HTTP_METHODS = Hash.new do |h, k|
-              # #to_s is required because user input could be symbol or string
-              h[k] = k.to_s.upcase
-            end
-            HTTP_METHODS_TO_SPAN_NAMES = Hash.new do |h, k|
-              h[k] = k.to_s
-              h[k] = 'HTTP' if k == '_OTHER'
-            end
-
             # Constant for the HTTP status range
             HTTP_STATUS_SUCCESS_RANGE = (100..399)
 
             def http_request(url, action_name, options = {})
-              @otel_method = ACTION_NAMES_TO_HTTP_METHODS[action_name]
+              @otel_method = action_name
               super
             end
 
@@ -39,10 +30,13 @@ module OpenTelemetry
               otel_before_request
               super
             rescue StandardError => e
-              # If an exception occurs before we can call `complete`, we should add and error status and close the span
+              # If an exception occurs before we can call `complete`
+              # we should add an error status and close the span
+              # and raise the original error
               @otel_span&.status = OpenTelemetry::Trace::Status.error("Request threw an exception: #{e.message}")
               @otel_span&.finish
               @otel_span = nil
+              raise e
             end
 
             def complete
@@ -73,12 +67,11 @@ module OpenTelemetry
             end
 
             def otel_before_request
-              method = '_OTHER' # Could be GET or not HTTP at all
-              method = @otel_method if instance_variable_defined?(:@otel_method) && !@otel_method.nil?
+              span_data = HttpHelper.span_attrs_for_stable(@otel_method)
 
               @otel_span = tracer.start_span(
-                HTTP_METHODS_TO_SPAN_NAMES[method],
-                attributes: span_creation_attributes(method),
+                span_data.span_name,
+                attributes: span_creation_attributes(span_data),
                 kind: :client
               )
 
@@ -95,10 +88,8 @@ module OpenTelemetry
 
             private
 
-            def span_creation_attributes(method)
-              instrumentation_attrs = {
-                'http.request.method' => method
-              }
+            def span_creation_attributes(span_data)
+              instrumentation_attrs = {}
 
               uri = _otel_cleanse_uri(url)
               if uri
@@ -108,9 +99,7 @@ module OpenTelemetry
 
               config = Ethon::Instrumentation.instance.config
               instrumentation_attrs['peer.service'] = config[:peer_service] if config[:peer_service]
-              instrumentation_attrs.merge!(
-                OpenTelemetry::Common::HTTP::ClientContext.attributes
-              )
+              instrumentation_attrs.merge!(span_data.attributes)
             end
 
             # Returns a URL string with userinfo removed.