+++ instrumentation/net_http/test/opentelemetry/instrumentation/net/http/dup/instrumentation_test.rb
@@ -23,6 +23,7 @@ describe OpenTelemetry::Instrumentation::Net::HTTP::Instrumentation do
     stub_request(:get, 'http://example.com/success?hello=there').to_return(status: 200)
     stub_request(:post, 'http://example.com/failure').to_return(status: 500)
     stub_request(:get, 'https://example.com/timeout').to_timeout
+    stub_request(:get, 'http://example.com/users/123').to_return(status: 200)
 
     # this is currently a noop but this will future proof the test
     @orig_propagation = OpenTelemetry.propagation
@@ -89,6 +90,37 @@ describe OpenTelemetry::Instrumentation::Net::HTTP::Instrumentation do
       assert_requested(
         :post,
         'http://example.com/failure',
+        headers: { 'Traceparent' => "00-#{span.hex_trace_id}-#{span.hex_span_id}-01" }
+      )
+    end
+
+    it 'after request with unknown http method' do
+      stub_request(:purge, 'http://example.com/purge').to_return(status: 200)
+      uri = URI('http://example.com/purge')
+      Net::HTTP.start(uri.host, uri.port) do |http|
+        http.request(Net::HTTP::Purge.new(uri))
+      end
+
+      _(exporter.finished_spans.size).must_equal 1
+      _(span.name).must_equal 'HTTP'
+      # old semantic conventions
+      _(span.attributes['http.method']).must_equal '_OTHER'
+      _(span.attributes['http.scheme']).must_equal 'http'
+      _(span.attributes['http.status_code']).must_equal 200
+      _(span.attributes['http.target']).must_equal '/purge'
+      _(span.attributes['net.peer.name']).must_equal 'example.com'
+      _(span.attributes['net.peer.port']).must_equal 80
+      # stable semantic conventions
+      _(span.attributes['http.request.method']).must_equal '_OTHER'
+      _(span.attributes['http.request.method_original']).must_equal 'PURGE'
+      _(span.attributes['url.scheme']).must_equal 'http'
+      _(span.attributes['http.response.status_code']).must_equal 200
+      _(span.attributes['url.path']).must_equal '/purge'
+      _(span.attributes['server.address']).must_equal 'example.com'
+      _(span.attributes['server.port']).must_equal 80
+      assert_requested(
+        :purge,
+        'http://example.com/purge',
         headers: { 'Traceparent' => "00-#{span.hex_trace_id}-#{span.hex_span_id}-01" }
       )
     end
@@ -348,6 +380,24 @@ describe OpenTelemetry::Instrumentation::Net::HTTP::Instrumentation do
       _(span.attributes['server.port']).must_equal(443)
     ensure
       WebMock.disable_net_connect!
+    end
+
+    it 'uses url.template in span name when present in client context' do
+      client_context_attrs = { 'url.template' => '/users/{id}' }
+      OpenTelemetry::Common::HTTP::ClientContext.with_attributes(client_context_attrs) do
+        Net::HTTP.get('example.com', '/users/123')
+      end
+
+      _(exporter.finished_spans.size).must_equal 1
+      _(span.name).must_equal 'GET /users/{id}'
+      _(span.attributes['http.method']).must_equal 'GET'
+      _(span.attributes['http.request.method']).must_equal 'GET'
+      _(span.attributes['url.template']).must_equal '/users/{id}'
+      assert_requested(
+        :get,
+        'http://example.com/users/123',
+        headers: { 'Traceparent' => "00-#{span.hex_trace_id}-#{span.hex_span_id}-01" }
+      )
     end
 
     it 'emits a "connect" span when connecting through an non-ssl proxy' do