+++ instrumentation/ruby_kafka/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History: opentelemetry-instrumentation-ruby_kafka
 
+### v0.24.0 / 2025-10-22
+
+* BREAKING CHANGE: Min Ruby Version 3.2
+
+* ADDED: Min Ruby Version 3.2
+
 ### v0.23.1 / 2025-09-30
 
 * FIXED: Min OTel Ruby API 1.7